# 문의 상황 검색 관리 기능 추가

+++ src/main/java/com/ads/main/service/FileService.java
@@ -20,12 +20,16 @@ public class FileService {
     @Value("${app.file.uploadPath}")
     private String UPLOAD_LOCATION;
 
-    public void move(FilesVo filesVo) {
-
+    public void move(FilesVo filesVo, String fileType) {
         try {
             if (filesVo != null && filesVo.getNewFile()) {
                 File tempFile = Paths.get(UPLOAD_TEMP_LOCATION + "/" + filesVo.getFileName()).toFile();
+
                 File destinationFile = Paths.get(UPLOAD_LOCATION + "/" + filesVo.getFileName()).toFile();
+
+                if (tempFile.exists() && destinationFile.exists()) {
+                    throw new RuntimeException(fileType + "을 다시 등록 해주세요.");
+                }
 
                 if ( !destinationFile.getParentFile().exists() ) {
                     try {
@@ -35,12 +39,14 @@ public class FileService {
                         e.getStackTrace();
                     }
                 }
-                Files.move(tempFile, destinationFile);
+
+                if ( !destinationFile.exists() ) {
+                    Files.move(tempFile, destinationFile);
+                }
             }
         } catch (Exception e) {
-            log.error("# 사업자 등록증 오류", e);
-            throw new RuntimeException("사업자 등록증 파일 이동 중 오류 발생.");
+            log.error("# 파일 이동 중 오류", e);
+            throw new RuntimeException(fileType + " 이동 중 오류 발생.");
         }
-
     }
 }
